UI:   -logs lazily load on scroll   -mech upgrade form shows all levels in between, fix bug where it perma hides build columns   -bold selected piece name   -research form now only shows immediate dependencies/unlocks   -mouse over now always shows distance from selected tile   -show base sustainability   -fix mangled Replace dialog text
Research Work:
  -add logic to get only immediate dependencies/unlocks or all for the whole tree
  -fix upgradeOnly chance weight bug using the wrong type
  -upgradeOnly techs that also unlock non-upgradeOnly tech don't have a penalty to chance weight on first research
  -reduce upgradeOnly chance mult
  -prevent negative nextAvg

Changes:
  -allow rebuilding full-health turrets and factories (if you want to randomize stats)
  -constructor can automatically start building factory/turret on research without upgrading
  -attack deviation:
    *MechSP upgrade can modify deviation
    *slightly increase average attack deviation (even more for turret)
    *clamp <1% down to 0

Bugs:
  -attacker could attack on build
  -attack partial-upkeep bug with integer division
  -always call Die() instead of just RemovePiece()

+++ WinFormsApp1/ClassLibrary1/Pieces/Killable.cs
@@ -68,7 +68,7 @@ namespace ClassLibrary1.Pieces
         {
             double pct = 1;
             if (HitsCur < damage)
-                pct = HitsCur / damage;
+                pct = HitsCur / (double)damage;
             this._hitsCur -= damage;
             this._shieldCur -= shieldDmg;
             if (this.Dead)
